+++ lib/tiny_png/shrink.rb
@@ -48,7 +48,7 @@ module TinyPng::Shrink
     # passes our quick checks, so let's fire off the request
     response = self.class.post(
       '/api/shrink',
-      :basic_auth => { :username => @options[:username], :password => @options[:api_key] },
+      :basic_auth => { :username => @options[:api_user], :password => @options[:api_key] },
       :body => current_content.force_encoding('BINARY')
     )